Avoid duplicate key warnings on key
weekDaysNames have an array of days in abbreviated letters, e.g. "S" for Sunday, and "S" for Saturday. React Native does not like that the key element have duplicates of the same key. Causes the "encountered two children with the same key" warning. Adding the index of the weekDaysNames array avoids duplicates, removing the warning.